Crate aws_smithy_async

Future utilities and runtime-agnostic abstractions for smithy-rs.

Async runtime specific code is abstracted behind async traits, and implementations are provided via feature flag. For now, only Tokio runtime implementations are provided.

Modules

future

Useful runtime-agnostic future implementations.

rt

Async runtime agnostic traits and implementations.

time

Time source abstraction to support WASM and testing

Macros

assert_elapsed

Given an Instant and a Duration, assert time elapsed since Instant is equal to Duration. This macro allows for a 5ms margin of error.