Module aws_smithy_async::future

Useful runtime-agnostic future implementations.

Modules

fn_stream

Utility to drive a stream with an async function and a channel.

never

Provides the Never future that never completes.

now_or_later

Provides the NowOrLater future with an explicit Now variant

rendezvous

Rendezvous channel implementation

timeout

Provides the Timeout future for adding a timeout to another future.