aws_smithy_async::future
Module pagination_stream

Provides types to support stream-like operations for paginators.

Modules

collect

Module to extend the functionality of types in patination_stream module to allow for collecting elements of the stream into collection.

fn_stream

Module to define utility to drive a stream with an async function and a channel.

Structs

PaginationStream

Stream specifically made to support paginators.

TryFlatMap

Utility wrapper to flatten paginated results