Struct aws_smithy_async::future::pagination_stream::PaginationStream

pub struct PaginationStream<Item>(/* private fields */);

Stream specifically made to support paginators.

PaginationStream provides two primary mechanisms for accessing stream of data.

With .next() (or try_next()):

let mut stream: PaginationStream<Page> = operation_to_yield_paginator();
while let Some(page) = stream.next().await {
    // process `page`
}

With .collect() (or try_collect()):

let mut stream: PaginationStream<Page> = operation_to_yield_paginator();
let result = stream.collect::<Vec<Page>>().await;

PaginationStream is implemented in terms of FnStream, but the latter is meant to be used internally and not by external users.

Implementations

impl<Item> PaginationStream<Item>

pub fn new(stream: FnStream<Item>) -> Self

Creates a PaginationStream from the given FnStream.

pub async fn next(&mut self) -> Option<Item>

Consumes and returns the next Item from this stream.

pub fn poll_next(&mut self, cx: &mut Context<'_>) -> Poll<Option<Item>>

Poll an item from the stream

pub async fn collect<T: Collectable<Item>>(self) -> T

Consumes this stream and gathers elements into a collection.

impl<T, E> PaginationStream<Result<T, E>>

pub async fn try_next(&mut self) -> Result<Option<T>, E>

Yields the next item in the stream or returns an error if an error is encountered.

pub async fn try_collect(self) -> Result<Vec<T>, E>

Convenience method for .collect::<Result<Vec<_>, _>().