Function aws_smithy_async::future::rendezvous::channel

pub fn channel<T>() -> (Sender<T>, Receiver<T>)

Create a new rendezvous channel

Rendezvous channels are equivalent to a channel with a 0-sized buffer: A sender cannot send until this is an active receiver waiting. This implementation uses a semaphore to record demand and coordinate with the receiver.